+++ crates/ruma-events/CHANGELOG.md
@@ -21,6 +21,9 @@ Improvements:
 - Add unstable support for [MSC4354](https://github.com/matrix-org/matrix-spec-proposals/pull/4354) Sticky Events behind the `unstable-msc4354` feature flag.
 - Add the `zeroize(mut self)` method on identifiers, which will call the `zeroize` crate.
 - Add `RoomMessageEventContent::thread` accessor.
+- Add experimental support for [MSC4495] (Selective Presence).
+
+[MSC4495]: https://github.com/matrix-org/matrix-spec-proposals/pull/4495
 
 Bug fixes: